change of approach for label storage - removed the messaging and went straight to localStorage after fetching
