+++ src/common/constants.js
@@ -8,7 +8,6 @@ export const actionType = {
   STORE_METADATA: 'Store Metadata',
   GET_METADATA: 'Get Metadata',
   QUERY_LABELS: 'Query Labels',
-  STORE_LABELS: 'Store Labels',
   GET_LABELS: 'Get Labels',
   LIGHTNING_METADATA: 'Lightning Metadata',
   VISUALFORCE_METADATA: 'VisualForce Metadata'
